% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stratpd.R
\name{plot.stratpd}
\alias{plot.stratpd}
\title{Plot stratified partial dependence functions}
\usage{
\method{plot}{stratpd}(x, ...)
}
\arguments{
\item{x}{AN object that inherits from class \code{"stratpd"}.}

\item{...}{Additional optional arguments to be passed onto
\code{\link[graphics]{plot}}.}
}
\description{
Extends R's generic \code{\link[graphics]{plot}} function to handle objects
that inherit from class \code{"stratpd"}.
}
