% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratpd.R
\name{stratpd}
\alias{stratpd}
\title{Stratified partial dependence}
\usage{
stratpd(X, y, feature_name = NULL, ...)
}
\arguments{
\item{X}{A matrix-like R object (e.g., a data frame or matrix) containing
ONLY the feature columns from the training data.}

\item{y}{A vector or matrix-like R object (e.g., a data frame or matrix)
containing ONLY the target values from the training data.}

\item{feature_name}{Character string specifying the feature of interest.}

\item{...}{Additional optional arguments.}
}
\description{
TBD.
}
